Switch to module.registerHooks() and inline the hook
`module.registerHooks()` (Node 22.15+) accepts the resolve hook as
a function on the main thread, removing the worker-thread split
that `module.register()` requires. The hook code (a string-keyed
alias for `node:test` / `node:assert` / `node:assert/strict`) is
small enough to live alongside the rest of `run-shim.ts`, so
`shim-hooks.ts` is folded in and deleted.

The Node version floor is unchanged in practice: strip-only TS for
the test path already pins us to 22.18+, comfortably above the
22.15 minimum that `registerHooks` introduced in.

The `registerHooks` API has carried `Stability: 1.2 - Release
candidate` since v22.15 and remains so on the v26 docs — i.e. the
surface has been frozen across v22 / v23 / v24 / v25 / v26 majors
without a breaking change. For dev-only test wiring this is an
acceptable risk; revisiting if Node ever lands a breaking change
on this API costs one targeted patch in the family.

Verified locally: real `node --test` 23/23, `make test-shim`
23/23 (now via registerHooks-aliased imports), headless-chromium
browser bundle 23/23.
